k8s-project/api/src/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.6
h11==0.14.0
hashring==1.5.1
idna==3.10
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic_core==2.27.1
python-memcached==1.62
redis==5.2.1
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
